♻️ (Refactor) - Reorganize some component in categorize folder
