+++ package.json
@@ -48,6 +48,7 @@
     "react": "^18.3.1",
     "react-day-picker": "8.10.1",
     "react-dom": "^18.3.1",
+    "react-dropzone": "^14.2.10",
     "react-hook-form": "^7.53.0",
     "react-hotkeys-hook": "^4.5.1",
     "sharp": "^0.33.5",
