+++ src/app/image-generator/_components/preview/components/NoiseBackground.tsx
