+++ src/app/image-generator/_components/sidebar/components/FormatSelect.tsx
