+++ src/app/image-generator/_components/sidebar/components/ImageInput.tsx
